Artifacts exported but still need to be cleaned up
