+++ spark-tensorflow/prepare/src/main/scala/com/google/cloud/ml/samples/criteo/CriteoPreprocessingApplication.scala
@@ -16,20 +16,41 @@
 
 package com.google.cloud.ml.samples.criteo
 
-import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.{DataFrame, SparkSession}
 import scopt.OptionParser
 
 
-case class ClargConfig(basePath: String = "",
-                       relativeInputPath: String = "",
-                       relativeOutputPath: String = "",
-                       relativeModelPath: String = "",
-                       numPartitions: Int = 500,
-                       loadModel: Boolean = false)
+/**
+  * Union of the different modes in which preprocessing can be done.
+  */
+sealed trait PreprocessingMode
+
+case object Analyze extends PreprocessingMode
+
+case object Transform extends PreprocessingMode
+
+/**
+  * Converts string for mode into the appropriate `PreprocessingMode` object.
+  */
+object PreprocessingMode {
+  def apply(specifier: String): Option[PreprocessingMode] = specifier.toLowerCase match {
+    case "analyze" => Some(Analyze)
+    case "transform" => Some(Transform)
+    case _ => None
+  }
+}
+
+
+case class NewClargConfig(basePath: String = "",
+                          relativeInputPath: String = "",
+                          relativeOutputPath: String = "",
+                          mode: PreprocessingMode = Analyze,
+                          numPartitions: Int = 500)
 
 object CriteoPreprocessingApplication {
+
   def main(args: Array[String]) {
-    val parser = new OptionParser[ClargConfig]("Criteo TFRecord Preprocessor") {
+    val parser = new OptionParser[NewClargConfig]("Criteo TFRecord Preprocessor") {
       head("CriteoPreprocessingApplication", "1.0.0")
 
       help("help").text("Prints this description of the CLI to the Criteo TFRecord Preprocessor")
@@ -43,23 +64,30 @@ object CriteoPreprocessingApplication {
       )
 
       opt[String]('o', "out").required.action((o, c) => c.copy(relativeOutputPath = o)).text(
-        "The relative path to the directory in which the resulting TFRecord files should be stored."
-      )
-
-      opt[String]('x', "indexer").required.action((x, c) => c.copy(relativeModelPath = x)).text(
-        "The path on which the model used to encode the input data either IS or SHOULD BE stored."
+        "The relative path to the directory in which the resulting transformed TFRecord files" +
+          " or analyze artifacts should be stored."
       )
 
       opt[Int]('n', "numPartitions").action((n, c) => c.copy(numPartitions = n)).text(
         "The number of partitions in which to process the input file. Default is 500."
       )
 
-      opt[Unit]('l', "load").action((l, c) => c.copy(loadModel = true)).text(
-        "Flag to denote whether model should be loaded from the given model path."
+      opt[String]('m', "mode").action(
+        (m, c) => {
+          val mod = PreprocessingMode(m)
+          c.copy(mode =
+            mod match {
+              case Some(mod) => mod
+              case None =>
+                throw new Exception("Illegal mode passed under -m or --mode." +
+                  "Pass \"train\", \"predict\", or \"evaluate\".")
+            })
+        }
+      ).text(
+        "\"analyze\", \"transform\""
       )
     }
-
-    parser.parse(args, ClargConfig()) match {
+    parser.parse(args, NewClargConfig()) match {
       case Some(config) =>
         implicit val spark = SparkSession.builder().
           appName("Criteo TFRecord Preprocessor").
@@ -68,46 +96,34 @@ object CriteoPreprocessingApplication {
         val inputPath = config.basePath ++ config.relativeInputPath
         val outputPath = config.basePath ++ config.relativeOutputPath
 
-        val modelPath = config.basePath ++ config.relativeModelPath
-    }
-  }
-}
-/*
         val features = CriteoFeatures(config.mode)
 
-        val importer = config.mode match {
-          //case Predict => new CleanTSVStreamImporter(inputPath, features.inputSchema)
-          case _ => new CleanTSVImporter(inputPath, features.inputSchema, config.numPartitions)
-        }
-
-        val exporter = config.mode match {
-          //case Predict => new FileStreamExporter(outputPath, "tfrecords")
-          case _ => new FileExporter(outputPath, "tfrecords")
-        }
-
         val artifactExporter = config.mode match {
-          //case Train => new FileArtifactExporter(config.basePath ++ "artifacts/")
+          case Analyze => new FileArtifactExporter(config.basePath ++ "artifacts/")
           case _ => new EmptyArtifactExporter()
         }
 
+        val indexer = new TrainingIndexer(features, artifactExporter)
+
+
+        if (config.mode == Analyze) {
+          val importer = new CleanTSVImporter(inputPath, features.inputSchema, config.numPartitions)
+          val missingReplacer = new CriteoMissingReplacer(artifactExporter)
+
+          val cleanedDf = importer.criteoImport
+          val averages = missingReplacer.getAverageIntegerFeatures(
+            cleanedDf, features.integerFeatureLabels)
+          averages.foreach {
+            case (col: String, df: DataFrame) =>
+              artifactExporter.export(col, df)
+          }
+
+          val embeddings = indexer.getEmbeddings(cleanedDf)
+          embeddings.foreach {
+            case (col: String, df: DataFrame) =>
+              artifactExporter.export(col, df)
+          }
+        }
     }
-*/
-        /*val indexer: CriteoIndexer = config.mode match {
-          case Train =>
-            val exporter = new FileExporter(modelPath, "csv")
-            new TrainingIndexer(features, artifactExporter)
-
-          case _ =>
-            val importer = new CleanTSVImporter(modelPath ++ "/*.csv", IndexerModelSchema.schema,
-              config.numPartitions)
-            new IndexApplier(features, importer)
-        } */
-
-        val missingReplacer = new CriteoMissingReplacer(artifactExporter)
-
-        val preprocessor =
-          new CriteoPreprocessor(config.mode, importer, indexer, exporter)
-        //preprocessor.execute()
-
-      case None => throw new Exception("Unable to configure preprocessor")
-    }*/
+  }
+}
